+++ config/kafka.php
@@ -59,11 +59,6 @@ return [
      */
     'debug' => env('KAFKA_DEBUG', false),
 
-    /*
-     | Repository for batching messages together
-     | Implement BatchRepositoryInterface to save batches in different storage
-     */
-    'batch_repository' => env('KAFKA_BATCH_REPOSITORY', \Junges\Kafka\BatchRepositories\InMemoryBatchRepository::class),
 
     /*
      | The sleep time in milliseconds that will be used when retrying flush