+++ src/Contracts/MessageProducer.php
@@ -2,8 +2,6 @@
 
 namespace Junges\Kafka\Contracts;
 
-use Junges\Kafka\Facades\Kafka;
-use Junges\Kafka\Producers\MessageBatch;
 use Junges\Kafka\Producers\Producer;
 use Junges\Kafka\Support\Testing\Fakes\ProducerFake;
 
@@ -68,12 +66,4 @@ interface MessageProducer extends InteractsWithConfigCallbacks
     public function send(): bool;
 
     public function build(): Producer|ProducerFake;
-
-    /**
-     * Send a message batch to Kafka.
-     *
-     * @throws \Junges\Kafka\Exceptions\CouldNotPublishMessage
-     * @deprecated Please use {@see Kafka::asyncPublish()} instead of batch messages.
-     */
-    public function sendBatch(MessageBatch $messageBatch): int;
 }