+++ src/Contracts/Producer.php
@@ -2,9 +2,6 @@
 
 namespace Junges\Kafka\Contracts;
 
-use Junges\Kafka\Exceptions\CouldNotPublishMessage;
-use Junges\Kafka\Producers\MessageBatch;
-
 interface Producer
 {
     /**
@@ -14,12 +11,6 @@ interface Producer
      * @throws \Exception
      */
     public function produce(ProducerMessage $message): bool;
-
-    /**
-     * @throws CouldNotPublishMessage
-     * @throws \Junges\Kafka\Exceptions\CouldNotPublishMessageBatch
-     */
-    public function produceBatch(MessageBatch $messageBatch): int;
 
     /**
      * @throws \Junges\Kafka\Exceptions\Transactions\TransactionShouldBeRetriedException