+++ src/Support/Testing/Fakes/ProducerFake.php
@@ -7,9 +7,6 @@ use Junges\Kafka\Concerns\ManagesTransactions;
 use Junges\Kafka\Config\Config;
 use Junges\Kafka\Contracts\Producer;
 use Junges\Kafka\Contracts\ProducerMessage;
-use Junges\Kafka\Exceptions\CouldNotPublishMessageBatch;
-use Junges\Kafka\Message\Message;
-use Junges\Kafka\Producers\MessageBatch;
 use RdKafka\Conf;
 
 class ProducerFake implements Producer
@@ -43,29 +40,6 @@ class ProducerFake implements Producer
         }
 
         return true;
-    }
-
-    /** @throws \Junges\Kafka\Exceptions\CouldNotPublishMessageBatch */
-    public function produceBatch(MessageBatch $messageBatch): int
-    {
-        if ($messageBatch->getTopicName() === '') {
-            throw CouldNotPublishMessageBatch::invalidTopicName($messageBatch->getTopicName());
-        }
-
-        $produced = 0;
-
-        if ($this->producerCallback !== null) {
-            $callback = $this->producerCallback;
-
-            /** @var Message $message */
-            foreach ($messageBatch->getMessages() as $message) {
-                $message->onTopic($messageBatch->getTopicName());
-                $callback($message);
-                $produced++;
-            }
-        }
-
-        return $produced;
     }
 
     public function flush(): int